requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
python-telegram-bot==20.7
python-dotenv==1.0.0
feedparser==6.0.10
lxml==4.9.3
Pillow==10.0.1 